fix(fsutil): process-unique case-probe names with collision retry (codex P1)
Two first-probe processes shared the constant probe path and EEXIST was
treated as case-insensitivity evidence for the lifetime of the loser,
producing wrong folding decisions on case-sensitive roots. Probe names are
now process-unique (pid + crypto-random + ordinal), collisions retry with
fresh entropy (8, bounded), and only a stat comparison against the file the
probe itself created decides case sensitivity. Read-only stat roots and
exact-path cleanup semantics preserved. Also removes the redundant
conversion flagged by unconvert on the linux pid helper.
Strict patch gate verified at 100% (1713/0/0); golangci-lint 0 issues.

+++ internal/fsutil/replacement_busy_pid_linux.go
@@ -45,7 +45,7 @@ func replacementProcessStartTimePlatform(pid int) *time.Time {
 	}
 	seconds := startTicks / clockTicks
 	remainingTicks := startTicks % clockTicks
-	if seconds > int64((1<<63-1)/int64(time.Second)) {
+	if seconds > int64((1<<63-1)/time.Second) {
 		return nil
 	}
 	start := time.Unix(bootSeconds, 0).
